+++ core/src/main/kotlin/net/corda/core/serialization/amqp/SerializerFactory.kt
@@ -163,13 +163,3 @@ class SerializerFactory {
                 Symbol::class.java to "symbol")
     }
 }
-
-class DeserializedGenericArrayType(private val componentType: Type) : GenericArrayType {
-    override fun getGenericComponentType(): Type = componentType
-    override fun getTypeName(): String = "${componentType.typeName}[]"
-    override fun toString(): String = typeName
-    override fun hashCode(): Int = componentType.hashCode() * 31
-    override fun equals(other: Any?): Boolean {
-        return other is GenericArrayType && componentType.equals(other.genericComponentType)
-    }
-}
